requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
pyspark>=3.5.0
delta-spark>=3.0.0
pytesseract>=0.3.10
opencv-python>=4.8.0
pyyaml>=6.0
tqdm>=4.65.0
matplotlib>=3.5.0
